requirements.txt
streamlit==1.28.1
pypdf2==3.0.1
langchain==0.0.350
openai==1.3.7
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
ruff==0.1.6
black==23.11.0
mypy==1.7.1
